fix(ci-tests): split the integration harness so the file-length gate passes
`make check-backend` fails on `main`: harness.go crossed the 500-line
whole-tree ceiling that `scripts/check-go-file-length.sh` holds, which
reds `main-health` and every branch that merges it.

The file had grown two concepts. The environment half — the migrated
database, the schema and owner pools, the seat and passport contexts,
River's schema — stays in harness.go. The fixture half — the id wideners
and every `Seed*` writer, plus the workspace-bound exec/count helpers
they use — moves to harnessseed.go beside it. Same package, so no caller
changes.

`MakeCapturePrivate`'s visibility guard now compares against the
`objPerson`/`objOrg` constants the package already declares; goconst
counts occurrences per file and the literals were its third use.
